+++ phantasm-at-rest/src/main/java/com/chiralbehaviors/CoRE/phantasm/authentication/AgencyBearerTokenAuthenticator.java
@@ -124,7 +124,7 @@ public class AgencyBearerTokenAuthenticator implements ModelAuthenticator,
         }
 
         // Validate agency has login cap to this core instance
-        if (!model.checkPermission(credential.roles.stream()
+        if (!model.checkExistentialPermission(credential.roles.stream()
                                                    .map(id -> model.records()
                                                                    .resolve(id))
                                                    .filter(a -> a != null)
