+++ phantasm-at-rest/src/main/java/com/chiralbehaviors/CoRE/phantasm/graphql/MetaContext.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright (c) 2017 Chiral Behaviors, LLC, all rights reserved.
+ * 
+ 
+ *  This file is part of Ultrastructure.
+ *
+ *  Ultrastructure is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  ULtrastructure is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with Ultrastructure.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.chiralbehaviors.CoRE.phantasm.graphql;
+
+import com.chiralbehaviors.CoRE.domain.Product;
+import com.chiralbehaviors.CoRE.meta.Model;
+import com.chiralbehaviors.CoRE.phantasm.graphql.mutations.ExistentialMutations;
+import com.chiralbehaviors.CoRE.phantasm.graphql.queries.ExistentialQueries;
+import com.chiralbehaviors.CoRE.phantasm.graphql.queries.WorkspaceQueries;
+
+/**
+ * @author halhildebrand
+ *
+ */
+public class MetaContext extends ExistentialContext
+        implements WorkspaceQueries {
+    public static interface MetaMutations extends ExistentialMutations {
+    }
+
+    public static interface MetaQueries
+            extends ExistentialQueries, WorkspaceQueries {
+    }
+
+    public MetaContext(Model model, Product workspace) {
+        super(model, workspace);
+    }
+
+}
