+++ phantasm-at-rest/src/main/java/com/chiralbehaviors/CoRE/phantasm/graphql/WorkspaceContext.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Chiral Behaviors, LLC, all rights reserved.
+ * Copyright (c) 2017 Chiral Behaviors, LLC, all rights reserved.
  *
 
  *  This file is part of Ultrastructure.
@@ -20,197 +20,154 @@
 
 package com.chiralbehaviors.CoRE.phantasm.graphql;
 
-import java.util.Collections;
+import static com.chiralbehaviors.CoRE.phantasm.graphql.WorkspaceSchema.ctx;
+
 import java.util.List;
-import java.util.Map;
-import java.util.Stack;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import javax.validation.constraints.NotNull;
 
-import com.chiralbehaviors.CoRE.domain.ExistentialRuleform;
+import com.chiralbehaviors.CoRE.domain.Agency;
 import com.chiralbehaviors.CoRE.domain.Product;
+import com.chiralbehaviors.CoRE.domain.Relationship;
+import com.chiralbehaviors.CoRE.jooq.tables.records.ExistentialRecord;
+import com.chiralbehaviors.CoRE.kernel.phantasm.CoreUser;
+import com.chiralbehaviors.CoRE.kernel.phantasm.Role;
 import com.chiralbehaviors.CoRE.meta.Model;
-import com.chiralbehaviors.CoRE.phantasm.graphql.WorkspaceSchema.MetaMutations;
-import com.chiralbehaviors.CoRE.phantasm.graphql.WorkspaceSchema.MetaQueries;
 import com.chiralbehaviors.CoRE.phantasm.graphql.WorkspaceSchema.Mutations;
 import com.chiralbehaviors.CoRE.phantasm.graphql.WorkspaceSchema.Queries;
-import com.chiralbehaviors.CoRE.phantasm.graphql.mutations.CoreUserAdmin;
-import com.chiralbehaviors.CoRE.phantasm.model.PhantasmCRUD;
-import com.chiralbehaviors.CoRE.phantasm.model.Phantasmagoria.NetworkAuthorization;
+import com.chiralbehaviors.CoRE.phantasm.graphql.types.Job;
+import com.chiralbehaviors.CoRE.phantasm.graphql.types.JobChronology;
 
-import graphql.ExceptionWhileDataFetching;
-import graphql.ExecutionResult;
-import graphql.GraphQL;
-import graphql.execution.ExecutionContext;
-import graphql.execution.ExecutionStrategy;
-import graphql.execution.SimpleExecutionStrategy;
-import graphql.language.Field;
+import graphql.annotations.GraphQLName;
 import graphql.schema.DataFetchingEnvironment;
-import graphql.schema.GraphQLFieldDefinition;
-import graphql.schema.GraphQLObjectType;
-import graphql.schema.GraphQLSchema;
-import graphql.schema.GraphQLType;
 
 /**
- * @author hhildebrand
+ * @author halhildebrand
  *
  */
-abstract public class WorkspaceContext extends PhantasmCRUD implements Queries,
-        Mutations, MetaQueries, MetaMutations, CoreUserAdmin {
-    public static class EdgeFieldBuilder
-            extends GraphQLFieldDefinition.Builder {
-        private NetworkAuthorization auth;
-
-        public EdgeFieldBuilder auth(NetworkAuthorization auth) {
-            this.auth = auth;
-            return this;
-        }
-
-        @Override
-        public GraphQLFieldDefinition build() {
-            return new GraphQLFieldEdgeDefinition(super.build(), auth);
-        }
-    }
-
-    public static class Traversal {
-        public final NetworkAuthorization auth;
-        public ExistentialRuleform        child;
-        public final ExistentialRuleform  parent;
-
-        public Traversal(ExistentialRuleform parent,
-                         NetworkAuthorization auth) {
-            this.parent = parent;
-            this.auth = auth;
-        }
-    }
-
-    private static class GraphQLFieldEdgeDefinition
-            extends GraphQLFieldDefinition {
-
-        private final NetworkAuthorization auth;
-
-        public GraphQLFieldEdgeDefinition(GraphQLFieldDefinition def,
-                                          NetworkAuthorization auth) {
-            super(def.getName(), def.getDescription(), def.getType(),
-                  def.getDataFetcher(), def.getArguments(),
-                  def.getDeprecationReason());
-            this.auth = auth;
-        }
-
-        public NetworkAuthorization getAuth() {
-            return auth;
-        }
-    }
-
-    private class SimpleTraversalStrategy extends SimpleExecutionStrategy {
-
-        @Override
-        protected ExecutionResult completeValue(ExecutionContext executionContext,
-                                                GraphQLType fieldType,
-                                                List<Field> fields,
-                                                Object result) {
-            if (!path.isEmpty() && fieldType instanceof GraphQLObjectType) {
-                Traversal edge = path.peek();
-                if (edge != null) {
-                    edge.child = (ExistentialRuleform) result;
-                }
-            }
-            return super.completeValue(executionContext, fieldType, fields,
-                                       result);
-        }
-
-        @Override
-        protected ExecutionResult resolveField(ExecutionContext executionContext,
-                                               GraphQLObjectType parentType,
-                                               Object source,
-                                               List<Field> fields) {
-            GraphQLFieldDefinition fieldDef = getFieldDef(executionContext.getGraphQLSchema(),
-                                                          parentType,
-                                                          fields.get(0));
-
-            Map<String, Object> argumentValues = valuesResolver.getArgumentValues(fieldDef.getArguments(),
-                                                                                  fields.get(0)
-                                                                                        .getArguments(),
-                                                                                  executionContext.getVariables());
-            DataFetchingEnvironment environment = new DataFetchingEnvironment(source,
-                                                                                  argumentValues,
-                                                                                  executionContext.getRoot(),
-                                                                                  fields,
-                                                                                  fieldDef.getType(),
-                                                                                  parentType,
-                                                                                  executionContext.getGraphQLSchema());
-
-            Object resolvedValue = null;
-            boolean pop = false;
-            try {
-                try {
-                    resolvedValue = fieldDef.getDataFetcher()
-                                            .get(environment);
-                    if (fieldDef instanceof GraphQLFieldEdgeDefinition) {
-                        path.push(new Traversal((ExistentialRuleform) source,
-                                                ((GraphQLFieldEdgeDefinition) fieldDef).getAuth()));
-                        pop = true;
-                    }
-                } catch (Exception e) {
-                    log.warn("Exception while fetching data", e);
-                    executionContext.addError(new ExceptionWhileDataFetching(e));
-                }
-
-                return completeValue(executionContext, fieldDef.getType(),
-                                     fields, resolvedValue);
-            } finally {
-                if (pop) {
-                    path.pop();
-                }
-            }
-        }
-    }
-
-    private static final Logger log = LoggerFactory.getLogger(WorkspaceContext.class);
-
-    public static Product getWorkspace(DataFetchingEnvironment env) {
-        return ((WorkspaceContext) env.getContext()).getWorkspace();
-    }
-
-    public static EdgeFieldBuilder newEdgeFieldDefinition() {
-        return new EdgeFieldBuilder();
-    }
-
-    private final Stack<Traversal> path = new Stack<>();
-    private final Product          workspace;
-
+public class WorkspaceContext extends ExistentialContext
+        implements Queries, Mutations {
     public WorkspaceContext(Model model, Product workspace) {
-        super(model);
-        this.workspace = workspace;
+        super(model, workspace);
     }
 
-    public ExecutionResult execute(GraphQLSchema schema, String query) {
-        return execute(schema, query, Collections.emptyMap());
+    @Override
+    public Job job(UUID id, DataFetchingEnvironment env) {
+        return new Job(Job.fetch(env, id));
     }
 
-    public ExecutionResult execute(GraphQLSchema schema, String query,
-                                   Map<String, Object> variables) {
-        ExecutionResult result = new GraphQL(schema,
-                                             getStrategy()).execute(query, this,
-                                                                    variables);
-        if (result.getErrors()
-                  .isEmpty()) {
-            return result;
+    @Override
+    public List<Job> jobs(List<UUID> ids, DataFetchingEnvironment env) {
+        return ids.stream()
+                  .map(id -> Job.fetch(env, id))
+                  .map(r -> new Job(r))
+                  .collect(Collectors.toList());
+    }
+
+    @Override
+    public Boolean authorized(UUID permission, UUID existential,
+                              DataFetchingEnvironment env) {
+        Model model = ctx(env);
+        return model.checkPermission(model.getCurrentPrincipal()
+                                          .getAsserted(),
+                                     (ExistentialRecord) model.records()
+                                                              .resolve(existential),
+                                     (Relationship) model.records()
+                                                         .resolve(permission));
+    }
+
+    @Override
+    public Boolean authorizedIfActive(UUID permission, UUID existential,
+                                      List<String> roleIds,
+                                      DataFetchingEnvironment env) {
+        Model model = ctx(env);
+        List<Agency> roleAgencies = roleIds.stream()
+                                           .map(s -> UUID.fromString(s))
+                                           .map(id -> model.records()
+                                                           .resolve(id))
+                                           .map(e -> (Agency) e)
+                                           .collect(Collectors.toList());
+        List<Role> roles = roleAgencies.stream()
+                                       .map(a -> model.wrap(Role.class, a))
+                                       .collect(Collectors.toList());
+        if (!model.wrap(CoreUser.class, model.getCurrentPrincipal()
+                                             .getPrincipal())
+                  .getRoles()
+                  .containsAll(roles)) {
+            return false;
         }
-        return result;
+        return model.checkExistentialPermission(roleAgencies, model.records()
+                                                                   .resolve(existential),
+                                                model.records()
+                                                     .resolve(permission));
     }
 
-    public Traversal getCurrentEdge() {
-        return path.isEmpty() ? null : path.peek();
+    @Override
+    public CoreUser getCurrentUser(DataFetchingEnvironment env) {
+        return ctx(env).wrap(CoreUser.class, ctx(env).getCurrentPrincipal()
+                                                     .getPrincipal());
     }
 
-    public Product getWorkspace() {
-        return workspace;
+    @Override
+    public Boolean hasRole(@NotNull @GraphQLName("role") String roleId,
+                           DataFetchingEnvironment env) {
+        Model model = ctx(env);
+        Role role = Optional.of(roleId)
+                            .map(s -> UUID.fromString(s))
+                            .map(id -> model.records()
+                                            .resolve(id))
+                            .map(e -> (Agency) e)
+                            .map(a -> model.wrap(Role.class, a))
+                            .get();
+        CoreUser authenticated = model.wrap(CoreUser.class,
+                                            model.getCurrentPrincipal()
+                                                 .getPrincipal());
+        return authenticated.getRoles()
+                            .contains(role);
     }
 
-    protected ExecutionStrategy getStrategy() {
-        return new SimpleTraversalStrategy();
+    @Override
+    public Boolean hasRoles(List<String> roleIds, DataFetchingEnvironment env) {
+        Model model = ctx(env);
+        List<Role> roles = roleIds.stream()
+                                  .map(s -> UUID.fromString(s))
+                                  .map(id -> model.records()
+                                                  .resolve(id))
+                                  .map(e -> (Agency) e)
+                                  .map(a -> model.wrap(Role.class, a))
+                                  .collect(Collectors.toList());
+        CoreUser authenticated = model.wrap(CoreUser.class,
+                                            model.getCurrentPrincipal()
+                                                 .getPrincipal());
+        return authenticated.getRoles()
+                            .containsAll(roles);
     }
 
+    @Override
+    public Boolean inRoles(List<String> roleIds, DataFetchingEnvironment env) {
+        Model model = ctx(env);
+        List<Agency> roles = roleIds.stream()
+                                    .map(s -> UUID.fromString(s))
+                                    .map(id -> model.records()
+                                                    .resolve(id))
+                                    .map(e -> (Agency) e)
+                                    .collect(Collectors.toList());
+        return model.getCurrentPrincipal()
+                    .getAsserted()
+                    .containsAll(roles);
+    }
+
+    public JobChronology jobChronology(UUID id, DataFetchingEnvironment env) {
+        return JobChronology.fetch(env, id);
+    }
+
+    public List<JobChronology> JobChronologies(List<UUID> ids,
+                                               DataFetchingEnvironment env) {
+        return ids.stream()
+                  .map(id -> JobChronology.fetch(env, id))
+                  .collect(Collectors.toList());
+    }
 }
