+++ phantasm-at-rest/src/main/java/com/chiralbehaviors/CoRE/phantasm/graphql/queries/JobChronologyQueries.java
@@ -22,7 +22,6 @@ package com.chiralbehaviors.CoRE.phantasm.graphql.queries;
 
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import javax.validation.constraints.NotNull;
 
@@ -35,21 +34,14 @@ import graphql.schema.DataFetchingEnvironment;
 /**
  * @author hhildebrand
  *
- */ 
+ */
 public interface JobChronologyQueries {
 
     @GraphQLField
-    default JobChronology jobChronology(@NotNull @GraphQLName("id") String id,
-                                        DataFetchingEnvironment env) {
-        return JobChronology.fetch(env, UUID.fromString(id));
-    }
+    public JobChronology jobChronology(@NotNull @GraphQLName("id") UUID id,
+                                       DataFetchingEnvironment env);
 
     @GraphQLField
-    default List<JobChronology> JobChronologies(@NotNull @GraphQLName("ids") List<String> ids,
-                                                DataFetchingEnvironment env) {
-        return ids.stream()
-                  .map(s -> UUID.fromString(s))
-                  .map(id -> JobChronology.fetch(env, id))
-                  .collect(Collectors.toList());
-    }
+    public List<JobChronology> JobChronologies(@NotNull @GraphQLName("ids") List<UUID> ids,
+                                               DataFetchingEnvironment env);
 }
