+++ phantasm-at-rest/src/test/java/com/chiralbehaviors/CoRE/phantasm/graphql/JobSchemaTest.java
@@ -193,7 +193,7 @@ public class JobSchemaTest extends AbstractModelTest {
 
     private ObjectNode execute(GraphQLSchema schema, String query,
                                Map<String, Object> variables) {
-        ExistentialContext context = new ExistentialContext(model, model.getKernel()
+        WorkspaceContext context = new WorkspaceContext(model, model.getKernel()
                                                                     .getKernelWorkspace());
         ExecutionResult execute = context.execute(schema, query, variables);
         assertTrue(execute.getErrors()
